golabels: refactor integration tests
