+++ interpreter/golabels/integrationtests/busy_nocgo.go
@@ -1,0 +1,10 @@
+//go:build nocgo && linux
+
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package integrationtests // import "go.opentelemetry.io/ebpf-profiler/interpreter/golabels/integrationtests"
+
+//go:noinline
+func busyFunc() {
+}
